refactor: rename `when-hover` to `hover-controller`
